refactor: page dirs
