+++ projects/portfolio/src/client/routes/about.tsx
@@ -1,6 +1,5 @@
 import { createFileRoute } from '@tanstack/react-router'
-
-import { About } from '#/client/components/pages/About'
+import { About } from '#/client/pages/About'
 
 export const Route = createFileRoute('/about')({
   component: RouteComponent,
